Fixed settings.logfile, disuse logging.msg
Since settings.logfile was broken, nobody seems to ever have used it.
My fix is pretty slow, as I reopen the logfile.
We should keep it open when switching to Python’s logging.

The log levels 5 and even 6 used to be used when calling logging.msg.
I replaced them with logging.debug in preparation of moving to python’s logging.

I also discovered that the _sim module is highly broken.
I fixed a bit superficial stuff,
but using anything that’s not tested will likely break